refactor: moved error class to errors.ts
