+++ src/api/init.ts
@@ -24,7 +24,8 @@ import { createRosettaMempoolRouter } from './routes/rosetta/mempool';
 import { createRosettaBlockRouter } from './routes/rosetta/block';
 import { createRosettaAccountRouter } from './routes/rosetta/account';
 import { createRosettaConstructionRouter } from './routes/rosetta/construction';
-import { isProdEnv, logError, logger, LogLevel, InvalidRequestError, waiter } from '../helpers';
+import { isProdEnv, logError, logger, LogLevel, waiter } from '../helpers';
+import { InvalidRequestError } from '../errors';
 import { createWsRpcRouter } from './routes/ws/ws-rpc';
 import { createSocketIORouter } from './routes/ws/socket-io';
 import { createBurnchainRouter } from './routes/burnchain';
