+++ src/api/routes/block.ts
@@ -4,13 +4,8 @@ import { BlockListResponse } from '@stacks/stacks-blockchain-api-types';
 
 import { DataStore } from '../../datastore/common';
 import { getBlockFromDataStore } from '../controllers/db-controller';
-import {
-  timeout,
-  waiter,
-  has0xPrefix,
-  InvalidRequestError,
-  InvalidRequestErrorType,
-} from '../../helpers';
+import { timeout, waiter, has0xPrefix } from '../../helpers';
+import { InvalidRequestError, InvalidRequestErrorType } from '../../errors';
 import { parseLimitQuery, parsePagingQueryInput } from '../pagination';
 import { getBlockHeightPathParam, validateRequestHexInput } from '../query-helpers';
 import { getChainTipCacheHandler, setChainTipCacheHeaders } from '../controllers/cache-controller';
